requirements.txt
torch
torchvision
torchaudio
scikit-learn
matplotlib
pandas
tqdm
pyyaml
Pillow
fastapi
uvicorn
wandb
seaborn
streamlit
python-multipart
requests
plotly
